fixed bug on button toggle
