+++ src/index.js
@@ -1,8 +1,11 @@
 
-// import weatherInfo from './getWeatherData';
+import weatherInfo from './getWeatherData';
 import displayBody from './body';
 import searchBar from './search';
+import toggleBtn from './toggleBtn';
+
 
 displayBody();
-// weatherInfo('lagos');
+weatherInfo('lagos');
 searchBar();
+toggleBtn();
